+++ src/main/java/org/websoso/WSSServer/feed/service/ReportService.java
@@ -5,7 +5,9 @@ import static org.websoso.WSSServer.domain.common.ReportedType.IMPERTINENCE;
 import static org.websoso.WSSServer.domain.common.ReportedType.SPOILER;
 import static org.websoso.WSSServer.exception.error.CustomCommentError.ALREADY_REPORTED_COMMENT;
 import static org.websoso.WSSServer.exception.error.CustomCommentError.COMMENT_NOT_FOUND;
+import static org.websoso.WSSServer.exception.error.CustomFeedError.ALREADY_REPORTED_FEED;
 import static org.websoso.WSSServer.exception.error.CustomFeedError.FEED_NOT_FOUND;
+import static org.websoso.WSSServer.exception.error.CustomFeedError.SELF_REPORT_NOT_ALLOWED;
 import static org.websoso.WSSServer.exception.error.CustomUserError.USER_NOT_FOUND;
 
 import lombok.RequiredArgsConstructor;
@@ -21,22 +23,24 @@ import org.websoso.WSSServer.exception.exception.CustomUserException;
 import org.websoso.WSSServer.feed.domain.Comment;
 import org.websoso.WSSServer.feed.domain.Feed;
 import org.websoso.WSSServer.feed.domain.ReportedComment;
+import org.websoso.WSSServer.feed.domain.ReportedFeed;
 import org.websoso.WSSServer.feed.repository.CommentRepository;
 import org.websoso.WSSServer.feed.repository.FeedRepository;
 import org.websoso.WSSServer.feed.repository.ReportedCommentRepository;
+import org.websoso.WSSServer.feed.repository.ReportedFeedRepository;
 import org.websoso.WSSServer.repository.UserRepository;
 import org.websoso.WSSServer.service.DiscordMessageClient;
 import org.websoso.WSSServer.service.MessageFormatter;
 
 @Service
 @RequiredArgsConstructor
-@Transactional
-public class ReportedCommentService {
+public class ReportService {
 
     private final ReportedCommentRepository reportedCommentRepository;
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
     private final FeedRepository feedRepository;
+    private final ReportedFeedRepository reportedFeedRepository;
     private final DiscordMessageClient discordMessageClient;
 
     @Transactional
@@ -78,5 +82,34 @@ public class ReportedCommentService {
     private Feed getFeedOrException(Long feedId) {
         return feedRepository.findById(feedId)
                 .orElseThrow(() -> new CustomFeedException(FEED_NOT_FOUND, "feed with the given id was not found"));
+    }
+
+    @Transactional
+    public void reportFeed(User user, Long feedId, ReportedType reportedType) {
+        Feed feed = getFeedOrException(feedId);
+
+        if (isUserFeedOwner(feed.getUser(), user)) {
+            throw new CustomFeedException(SELF_REPORT_NOT_ALLOWED, "cannot report own feed");
+        }
+
+        if (reportedFeedRepository.existsByFeedAndUserAndReportedType(feed, user, reportedType)) {
+            throw new CustomFeedException(ALREADY_REPORTED_FEED, "feed has already been reported by the user");
+        }
+
+        reportedFeedRepository.save(ReportedFeed.create(feed, user, reportedType));
+
+        int reportedCount = reportedFeedRepository.countByFeedAndReportedType(feed, reportedType);
+        boolean shouldHide = reportedType.isExceedingLimit(reportedCount);
+
+        if (shouldHide) {
+            feed.hideFeed();
+        }
+
+        discordMessageClient.sendDiscordWebhookMessage(DiscordWebhookMessage.of(
+                MessageFormatter.formatFeedReportMessage(user, feed, reportedType, reportedCount, shouldHide), REPORT));
+    }
+
+    private Boolean isUserFeedOwner(User createdUser, User user) {
+        return createdUser.equals(user);
     }
 }
